backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP client
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3

# Optional: For async file operations (if needed)
aiofiles==23.2.1

